requirements.txt
javalang==0.13.0
gitpython==3.1.40
networkx>=3.2
python-louvain==0.16
leidenalg>=0.10.0
python-igraph>=0.11.0
anthropic>=0.40.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pytest>=7.4.0
pyyaml>=6.0.0
scikit-learn>=1.3.0
# Data science and ML
scipy>=1.11.0

# Git integration for Stage 7
GitPython>=3.1.40
